refactor(wizard): #1544 stage 1 — extract create_course graph guard
First commit of the 8-stage split per #1544. Pure shape change, no
behaviour change. The dispatcher tests at
`tests/lib/chat/wizard-tool-executor-dispatcher.test.ts` ("create_course
hard-fails when graph is incomplete") still pin the same hard-fail
contract.

## What moved

- `lib/chat/wizard-tool-executor/tools/create_course/_graph-guard.ts` (new)
  — exports `runGraphGuard(setupData)` returning
  `{ earlyReturn: WizardToolExec | null }`. `earlyReturn === null`
  means continue; a payload means caller returns it verbatim.
- `lib/chat/wizard-tool-executor/tools/create_course.ts` L13-35 replaced
  with a 3-line call site (1340 → 1321 lines on the orchestrator).

The dispatcher tests verify the return shape (`is_error: true`,
`missingKeys`, `missingLabels`) is byte-identical to the pre-extract
version. Zero new tsc / lint warnings.

## Sequencing — stages 2-8 still to extract

Per #1544's brief, remaining stages and their realistic boundaries on
the current file (post-#1545 = 1321 LOC):

| # | Stage | Approx range | Notes |
|---|---|---|---|
| 2 | resolveDomainOrError | L28-96 | validUuid cascade + slug/name + ensureInstitutionAndDomain + final guard; pulls in `courseName` extraction at L53 inline |
| 3 | resolveSubjectOrError | L55-116 | placeholder discipline guard + subjectDiscipline assignment |
| 4 | reuseExistingCoursePath | L114-498 | the big one — draftPlaybookId branch (~380 LOC); already calls `createPedagogyAssertionsFromCourseRef` from #1545 |
| 5 | newCourseScaffoldPath | L500-1015 | fresh-scaffold branch (~510 LOC); needs the new-path test pin |
| 6 | applyBehaviorTargetsBlock | inside L5 | not a separate stage at this file shape — folds into the new-path extract |
| 7 | enrollAndCreateCaller | L1017-1080ish | enrollCaller + randomFakeName + skipOnboarding wiring |
| 8 | generateLessonPlanAndReturn | L1081-1321 | lesson plan + final response payload |

Stage 4 + 5 are the bulk of the LOC and the riskiest cuts (shared
state via closed-over local variables: `domainId`, `playbookId`,
`subject`, `existingPlaybookId`, `caller`, `lessonPlanModel`, etc).
The BA brief calls for a shared `CreateCourseContext` interface
threaded through each stage helper. Subsequent commits introduce
the context shape one stage at a time so the dispatcher tests
(and the new-path pin from #1544 AC) can run between every commit.

## Out of scope here

- The new-path branch test pin (acceptance criterion of #1544). Mock
  surface is still too wide at this commit; the pin lands together
  with stage 5 (`_new-path.ts`) so the mock can be scoped to that
  file.
